feat(create): migrate playground to svelteConfig and require @sveltejs/kit ^2.62.0
The inlined vite.config config needs kit >= 2.62.0; snapshots regenerated.
