--- packages/sv/src/cli/tests/snapshots/create-only/vite.config.ts
@@ -1,6 +1,0 @@
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vite';
-
-export default defineConfig({
-	plugins: [sveltekit()]
-});
