--- packages/sv/src/cli/tests/snapshots/create-with-all-addons/e2e/demo.test.ts
@@ -1,6 +1,0 @@
-import { expect, test } from '@playwright/test';
-
-test('home page has expected h1', async ({ page }) => {
-	await page.goto('/');
-	await expect(page.locator('h1')).toBeVisible();
-});
