--- packages/sv/src/cli/tests/snapshots/create-with-all-addons/svelte.config.js
@@ -1,23 +1,0 @@
-import { mdsvex } from 'mdsvex';
-import adapter from '@sveltejs/adapter-node';
-
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	compilerOptions: {
-		// Force runes mode for the project, except for libraries. Can be removed in svelte 6.
-		runes: ({ filename }) => filename.split(/[/\\]/).includes('node_modules') ? undefined : true
-	},
-	kit: {
-		adapter: adapter(),
-		typescript: {
-			config: (config) => ({
-				...config,
-				include: [...config.include, '../drizzle.config.ts']
-			})
-		}
-	},
-	preprocess: [mdsvex({ extensions: ['.svx', '.md'] })],
-	extensions: ['.svelte', '.svx', '.md']
-};
-
-export default config;
