+++ packages/sv/src/cli/tests/snapshots/create-with-all-addons/vite.config.js
@@ -1,20 +1,36 @@
 import { paraglideVitePlugin } from '@inlang/paraglide-js';
+import { mdsvex } from 'mdsvex';
 import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'vitest/config';
 import { playwright } from '@vitest/browser-playwright';
+import adapter from '@sveltejs/adapter-node';
 import { sveltekit } from '@sveltejs/kit/vite';
 
 export default defineConfig({
 	plugins: [
 		tailwindcss(),
-		sveltekit(),
+		sveltekit({
+			compilerOptions: {
+				// Force runes mode for the project, except for libraries. Can be removed in svelte 6.
+				runes: ({ filename }) => filename.split(/[/\\]/).includes('node_modules') ? undefined : true
+			},
+			adapter: adapter(),
+			preprocess: [mdsvex({ extensions: ['.svx', '.md'] })],
+			extensions: ['.svelte', '.svx', '.md'],
+			typescript: {
+				config: (config) => ({
+					...config,
+					include: [...config.include, '../drizzle.config.ts']
+				})
+			}
+		}),
 		paraglideVitePlugin({ project: './project.inlang', outdir: './src/lib/paraglide' })
 	],
 	test: {
 		expect: { requireAssertions: true },
 		projects: [
 			{
-				extends: './vite.config.ts',
+				extends: './vite.config.js',
 				test: {
 					name: 'client',
 					browser: {
@@ -28,7 +44,7 @@ export default defineConfig({
 			},
 
 			{
-				extends: './vite.config.ts',
+				extends: './vite.config.js',
 				test: {
 					name: 'server',
 					environment: 'node',
