+++ packages/sv/src/create/playground.ts
@@ -4,6 +4,7 @@ import {
 	js,
 	parse,
 	svelte,
+	svelteConfig,
 	downloadJson,
 	Walker
 } from '@sveltejs/sv-utils';
@@ -255,12 +256,25 @@ export function setupPlaygroundProject(
 
 	let experimentalAsyncNeeded = true;
 	const addExperimentalAsync = () => {
-		const svelteConfigPath = path.join(cwd, filePaths.svelteConfig);
-		const svelteConfig = fs.readFileSync(svelteConfigPath, 'utf-8');
-		const { ast, generateCode } = parse.script(svelteConfig);
-		const { value: config } = js.exports.createDefault(ast, { fallback: js.object.create({}) });
-		js.object.overrideProperties(config, { compilerOptions: { experimental: { async: true } } });
-		fs.writeFileSync(svelteConfigPath, generateCode(), 'utf-8');
+		// `compilerOptions` is a svelte-level option, so it goes on `config` regardless of whether
+		// the config lives in `svelte.config.js` or inside `sveltekit()` in `vite.config.js`.
+		svelteConfig.edit(
+			{
+				cwd,
+				sv: {
+					file: (p, edit) => {
+						const full = path.join(cwd, p);
+						const content = fs.existsSync(full) ? fs.readFileSync(full, 'utf-8') : '';
+						const result = edit(content);
+						if (result === false || result === '') return;
+						fs.writeFileSync(full, result, 'utf-8');
+					}
+				}
+			},
+			({ override }) => {
+				override({ compilerOptions: { experimental: { async: true } } });
+			}
+		);
 	};
 
 	// we want to change the svelte version, even if the user decided
